GPS watch example updated for v3

+++ examples/GPS/GPS_Watch/src/main/java/gui/WatchGui.java
@@ -5,9 +5,12 @@ import tracking.shared.Unit;
 
 public interface WatchGui {
 
-    public void setData(float value, Unit unit);
+    public void setData(double value, Unit unit);
+
     public void setTime(int time);
+
     public void setIndicator(Indicator value);
+
     public void display();
-    
+
 }
